Wired the transaction viewmodel to the UI and Converted the Onboarding viewmodel from manual to Hilt

+++ app/src/main/java/com/amanansari/spendly/onBoarding/state/UserInfoUiState.kt
@@ -8,7 +8,7 @@ import com.amanansari.spendly.onBoarding.viewmodel.UserInfoStep
 data class UserInfoUiState(
     val name: String = "",
     val email: String = "",
-    val initialAmount: Double = 0.0,
+    val initialAmount: Long = 0L,
     val amountFieldValue: TextFieldValue = TextFieldValue(""),
     val currentUserInfoStep: UserInfoStep = UserInfoStep.NAME,
     val currency: CurrencyInfo = CurrencyInfo("INR", "₹")
